refactor(analysis): add language enum (#3994)
