+++ analysis/import/CoverageImporter/src/test/kotlin/CoverageImporterTest.kt
@@ -1,8 +1,10 @@
 import de.maibornwolff.codecharta.importer.coverage.CoverageImporter
+import de.maibornwolff.codecharta.importer.coverage.languages.Language
 import de.maibornwolff.codecharta.model.AttributeType
 import de.maibornwolff.codecharta.serialization.ProjectDeserializer
 import io.mockk.unmockkAll
 import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
 import org.skyscreamer.jsonassert.JSONParser
@@ -10,6 +12,7 @@ import picocli.CommandLine
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
 import java.io.File
+import java.io.FileNotFoundException
 import java.io.PrintStream
 
 class CoverageImporterTest {
@@ -212,6 +215,70 @@ class CoverageImporterTest {
         )
 
         assertThat(cliResult).contains(listOf("checksum", "data", "\"projectName\":\"\""))
+    }
+
+    @Test
+    fun `should find the default report file`() {
+        val directory = File("src/test/resources/languages/javascript/nested")
+        val expectedFile = File("src/test/resources/languages/javascript/nested/lcov.info")
+
+        val result = CoverageImporter().getReportFileFromString(directory.path, Language.JAVASCRIPT)
+
+        assertThat(result).isEqualTo(expectedFile)
+    }
+
+    @Test
+    fun `should return file if input is a file matching the default file`() {
+        val file = File("src/test/resources/languages/javascript/lcov.info")
+
+        val result = CoverageImporter().getReportFileFromString(file.path, Language.JAVASCRIPT)
+
+        assertThat(result).isEqualTo(file)
+    }
+
+    @Test
+    fun `should return file if input is a file with correct extension`() {
+        val file = File("src/test/resources/languages/javascript/minimal_lcov.info")
+
+        val result = CoverageImporter().getReportFileFromString(file.path, Language.JAVASCRIPT)
+
+        assertThat(result).isEqualTo(file)
+    }
+
+    @Test
+    fun `should throw no file found if there is no matching file`() {
+        val directory = File("src/test/kotlin")
+
+        assertThatThrownBy { CoverageImporter().getReportFileFromString(directory.path, Language.JAVASCRIPT) }
+            .isInstanceOf(FileNotFoundException::class.java)
+            .hasMessageContaining("No files matching lcov.info found in directory:")
+    }
+
+    @Test
+    fun `should throw file not found if file does not exist`() {
+        val file = File("src/test/resources/languages/javascript/non_existent_file.info")
+
+        assertThatThrownBy { CoverageImporter().getReportFileFromString(file.path, Language.JAVASCRIPT) }
+            .isInstanceOf(FileNotFoundException::class.java)
+            .hasMessageContaining("File not found")
+    }
+
+    @Test
+    fun `should throw no matching file extension if file does not match any known file extension`() {
+        val file = File("src/test/resources/languages/javascript/invalid_existing_file.txt")
+
+        assertThatThrownBy { CoverageImporter().getReportFileFromString(file.path, Language.JAVASCRIPT) }
+            .isInstanceOf(FileNotFoundException::class.java)
+            .hasMessageContaining("does not match any known file extension")
+    }
+
+    @Test
+    fun `should throw multiple files found if multiple files match the default file`() {
+        val directory = File("src/test/resources/languages/javascript")
+
+        assertThatThrownBy { CoverageImporter().getReportFileFromString(directory.path, Language.JAVASCRIPT) }
+            .isInstanceOf(FileNotFoundException::class.java)
+            .hasMessageContaining("Multiple files matching lcov.info found in directory:")
     }
 }
